New Game, Save Game logic
I worked on developing the logic for a new game and implemented the save game feature. Additionally, I created models to define the initial values of the cards, integrated a tournament counter, and established all the necessary URLs for these functionalities.
Redesigned the appearance of the start page when accessed through the lobby and upon the initial visit to the website.